moves logic to a single method + reuses for events
