+++ webapp/packages/core-events/src/ExceptionsCatcherService.ts
@@ -1,6 +1,6 @@
 /*
  * CloudBeaver - Cloud Database Manager
- * Copyright (C) 2020-2024 DBeaver Corp and others
+ * Copyright (C) 2020-2026 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0.
  * you may not use this file except in compliance with the License.
@@ -35,7 +35,8 @@ export class ExceptionsCatcherService extends Bootstrap {
 
   private readonly catcher = (event: Event | string, source?: string, lineno?: number, colno?: number, _error?: Error) => {
     if (_error) {
-      this.notificationService.logException(_error, this.messageTitle);
+      this.notificationService.logException(_error, 
+      _error.name || this.messageTitle);
     } else {
       const message = String(event);
       this.notificationService.logError({
